Move `SceneLibrary` implementation to its own file
